rename analyze_project file names
