+++ packages/flutter_tools/lib/src/commands/validate_project.dart
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import '../analyze_project.dart';
-import '../analyze_project_validator.dart';
+import '../project_validator_result.dart';
+import '../project_validator.dart';
 import '../base/file_system.dart';
 import '../base/logger.dart';
 import '../project.dart';
