+++ packages/flutter_tools/test/general.shard/analyze_project_test.dart
@@ -3,8 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:file/memory.dart';
-import 'package:flutter_tools/src/analyze_project.dart';
-import 'package:flutter_tools/src/analyze_project_validator.dart';
+import 'package:flutter_tools/src/project_validator_result.dart';
+import 'package:flutter_tools/src/project_validator.dart';
 import 'package:flutter_tools/src/project.dart';
 
 import '../src/common.dart';
